break dependency of main.init.buildgen on runner
