+++ core/internal/src/mill/internal/Util.scala
@@ -1,6 +1,63 @@
 package mill.internal
 
+import scala.reflect.NameTransformer.encode
+
 private[mill] object Util {
+
+  val alphaKeywords: Set[String] = Set(
+    "abstract",
+    "case",
+    "catch",
+    "class",
+    "def",
+    "do",
+    "else",
+    "enum",
+    "export",
+    "extends",
+    "false",
+    "final",
+    "finally",
+    "forSome",
+    "for",
+    "given",
+    "if",
+    "implicit",
+    "import",
+    "lazy",
+    "match",
+    "new",
+    "null",
+    "object",
+    "override",
+    "package",
+    "private",
+    "protected",
+    "return",
+    "sealed",
+    "super",
+    "then",
+    "this",
+    "throw",
+    "trait",
+    "try",
+    "true",
+    "type",
+    "val",
+    "var",
+    "while",
+    "with",
+    "yield",
+    "_",
+    "macro"
+  )
+
+  def backtickWrap(s: String): String = s match {
+    case s"`$v`" => s
+    case _ => if (encode(s) == s && !alphaKeywords.contains(s)) s
+    else "`" + s + "`"
+  }
+
   def leftPad(s: String, targetLength: Int, char: Char): String = {
     char.toString * (targetLength - s.length) + s
   }
