+++ runner/src/mill/runner/FileImportGraph.scala
@@ -7,7 +7,7 @@ import mill.runner.worker.api.{ImportTree, MillScalaParser}
 import scala.reflect.NameTransformer.encode
 import scala.collection.mutable
 import scala.jdk.CollectionConverters.CollectionHasAsScala
-
+import mill.internal.Util.backtickWrap
 /**
  * @param seenScripts
  * @param repos
@@ -33,59 +33,6 @@ case class FileImportGraph(
 @internal
 object FileImportGraph {
 
-  val alphaKeywords: Set[String] = Set(
-    "abstract",
-    "case",
-    "catch",
-    "class",
-    "def",
-    "do",
-    "else",
-    "enum",
-    "export",
-    "extends",
-    "false",
-    "final",
-    "finally",
-    "forSome",
-    "for",
-    "given",
-    "if",
-    "implicit",
-    "import",
-    "lazy",
-    "match",
-    "new",
-    "null",
-    "object",
-    "override",
-    "package",
-    "private",
-    "protected",
-    "return",
-    "sealed",
-    "super",
-    "then",
-    "this",
-    "throw",
-    "trait",
-    "try",
-    "true",
-    "type",
-    "val",
-    "var",
-    "while",
-    "with",
-    "yield",
-    "_",
-    "macro"
-  )
-
-  def backtickWrap(s: String): String = s match {
-    case s"`$v`" => s
-    case _ => if (encode(s) == s && !alphaKeywords.contains(s)) s
-      else "`" + s + "`"
-  }
 
   import mill.api.JsonFormatters.pathReadWrite
   implicit val readWriter: upickle.default.ReadWriter[FileImportGraph] = upickle.default.macroRW
